style(commits): add commitizen configuration
Adopt the setup that has been initially tested in datalad-next.

This setup is incomplete. It does not (yet) support the question-driven
commit generation via `cz`. However, given that this is not envisioned
to be a common approach the present setup should suffice.

Documentation on the setup will come in a later commit.

Refs: https://github.com/datalad/datasalad/issues/10