doc(changelog): remove UTF symbols from pyproject.toml
The reason is that `pyproject.toml` is read in many contexts. These
symbols are niceties that nevertheless impact platform compatibility.
Reading the file with the default platform encoding on windows
(`cp1252`) fails, because these symbols cannot be mapped.

Software should be able to handle UTF, but we cannot afford to make a
point in this crucial file.