ci(typeannotation): GitHub action to the whole project
This amends the `mypy-pr` workflow and runs a full check on the entire
project on `push`. We may want to have the PR-variant be a critical
check to pass, and this additional one be an information items of
an (unrelated) problem that needs fixing.

Rationale is that Python version upgrades, and mypy improvements may
introduce test failures at any time, and we cannot ask individual
PR authors to deal with them immediately for code that is unrelated
to a PR.

Closes: https://github.com/datalad/datasalad/issues/12