rf(CommandError): discontinue rendering of `stdout|stderr`
In its previous implementation, `CommandError.__str__()` could output
large amounts of test, due to the inclusion of provided `stdout` or
`stderr` as `str` or `bytes`. This is not compliant with the idea of
that function:

    compute the “informal” or nicely printable string representation
    of an object

This change removes the rendering of captured outputs. A future change
may re-add support for it with additional safe-guard on maximum length,
or rendering a summary (e.g., in the case of byte-strings) rather than
risking the rendering of large blobs of (human-)unreadable output.