rf(CommandError): make the `cmd` command mandatory
It makes little sense to report an exception on a command execution
failure without including information on the command that failed.

It is safe to assume that we know under all circumstances which command
was executed that lead to the failure. If we do not know this, the use
of `CommandError` is likely inappropriate