feat(CommandError): implement `__repr__()`
Without this implementation ``CommandError` would use ``RuntimeError``'s
variant and ignore all additional structured information except for
``.msg`` -- which is frequently empty and confuses with a
`CommandError('')` display.

This implementation limits the rendering output by truncating captured
``stdout`` and ``stderr``. For ``str``-type outputs the start and end
of the string is shown, with an indication how much text was cut in the
middle. For ``bytes`` an indication of the number of bytes is given,
only.

The handling of byte-strings is different from the one implemented in
datalad-core for this purpose. There, it tries to force-decode bytes to
unicode, while guessing the encoding, and raising an exception if that
is not possible. I consider it inappropriate to raise an exception
while handling a ``CommandError`` exception, because any client code
will need to deal with this complication. Moreover, it is not clear
that such decoding is possible or even meaningful in the general case
(the output could well be an actual data byte-stream, with no meaningful
human-readable rendering).

The existing test for exception rendering has been extended to cover
both ``str()`` and ``repr()``.

This change obsoletes the ``commanderror`` patch in datalad-next.

Refs: https://github.com/datalad/datalad-next/blob/cee9c8e2565555966e2b32c3ffb0eb39a4a7b86a/datalad_next/patches/commanderror.py