feat(CommandError): more readable `str()` rendering with signal name decoding
This change aligns the string rendering of `CommandError` with that of
`subprocess.CalledProcessError`.

`CommandError` reports more information, but for the matching subset
style and wording are aligned. This should give users a more familiar
behavior.

Moreover, this change also adopts the decoding of signal names done
in subprocess. For example, translating -6 to SIGABRT. In the case where
a signal report carries useful information, this could save someone a
search query