test(CommandError): document and verify amending `CommandError.msg`
Typically, the machinery that executes a command and raises a
`CommandError` is not aware of why a (failed) command was executed,
or whether there are probably causes for a command failure.

This change tests and documents a pattern that allows for adding
a context message to a `CommandError` instance.

Such messages can be used to help users sort out errors.