feat: `iterable_subprocess()` utility
This is a code import from datalad-next@0f4829f4b69b826057c2881f0318329480cc9f38

This 3rd-party implementation had been originally introduced in
datalad-next V1.1. It is a context manager that feeds input to
subprocesses via iterables, and also exposes their output as an
iterable. The implementation is based on
https://github.com/uktrade/iterable-subprocess, and this fork
has been modified to work homogeneously on the Windows platform
too.

This code is the foundation for a future code import of
`iter_subproc()`, the core subprocess execution utility of
datalad-next.

For documentation/compliance purposes, the code import include the
original license, README, and `pyproject.toml` metadata.