feat(CommandError): report `stderr` in `__str__()`
A truncated string with the error output is included, in an effort to
relay possible error messages. For byte strings an attempt is made to
decode it using the system's default encoding (see rationale in the
comments on why that and why not more).

The tests are extended with a few illustrative cases.

This change should provide the necessary error reporting behavior for
the `iterable_subprocess` tests to pass.