feat: `itertools` collection
This is a code import from datalad-next@067fa7c96b9306ee6a7b5ab16dc7aaea8591a7e2

This import brings a collection of iterators that are focused on
processing (byte) streams of information collected from subprocesses.
The can align and decode such stream in a way that eases the
implementation of consuming code that expects a particular protocol,
regardless of lower-level eventualities, such as variable chunking
of a bytes stream downloaded from a remote source.

Moreover, the import contains basic tooling to set up and run simple
pipe-like processing stream, including splitting off, and rejoining
certain items in such a stream.

The implementation aims to be usable in conjunction with standard Python
utilities for iterables, such as the built-in `itertools`, and the
`more_itertools` package.

Closes: https://github.com/datalad/datasalad/issues/25