fix(`decode_bytes`): handle multiple errors
This commit fixes a bug that was detected
by @mih. Due to this bug, multiple encoding
errors in a single input chunk of `decode_bytes`
lead to unexpected decoding-exceptions.

The commit also adds a regression test to enure
that multiple encoding errors in a single input
chunk are handled properly.