fix(`decode_bytes`): error handling led to data loss in subsequent chunks
The contained new test documents how error handling in `decode_bytes()`
caused data in a subsequent chunk to be skipped. The reason for that
was that the pointer variable was not reset when a chunk (or rather
the joined data of any previous chunks) was fully decoded. Decoding
in the next chunk would start from the position last recorded during
error handling in a previous chunk.

Refs: https://github.com/datalad/datalad-next/pull/744
Closes: https://github.com/datalad/datasalad/issues/38